requirements.txt
opencv-python>=4.8.0
imutils>=0.5.4
pygame>=2.5.0
pillow>=10.0.0
discord.py>=2.3.2
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6 